refactor(ecoletor): renomeia rota para UserEcoletorRoutes para manter padrao
